+++ src/test/groovy/com/github/gradle/node/npm/task/NpmTaskTest.groovy
@@ -15,11 +15,9 @@ class NpmTaskTest extends AbstractTaskTest {
 
     def "exec npm task"() {
         given:
-        props.setProperty('os.name', 'Linux')
         nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
 
         def task = project.tasks.create('simple', NpmTask)
-        mockPlatformHelper(task)
         mockProjectApiHelperExec(task)
         task.args.set(['a', 'b'])
         task.environment.set(['a': '1'])
@@ -39,11 +37,9 @@ class NpmTaskTest extends AbstractTaskTest {
 
     def "exec npm task (windows)"() {
         given:
-        props.setProperty('os.name', 'Windows')
         nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Windows", "x86_64", {}))
 
         def task = project.tasks.create('simple', NpmTask)
-        mockPlatformHelper(task)
         mockProjectApiHelperExec(task)
         task.args.set(['a', 'b'])
         task.environment.set(['a': '1'])
@@ -63,14 +59,12 @@ class NpmTaskTest extends AbstractTaskTest {
 
     def "exec npm task (download)"() {
         given:
-        props.setProperty('os.name', 'Linux')
         nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
         nodeExtension.download.set(true)
         def nodeDir = projectDir.toPath().resolve(".gradle").resolve("nodejs")
                 .resolve("node-v${DEFAULT_NODE_VERSION}-linux-x64")
 
         def task = project.tasks.create('simple', NpmTask)
-        mockPlatformHelper(task)
         mockProjectApiHelperExec(task)
         task.args.set(["run", "command"])
 
@@ -94,7 +88,6 @@ class NpmTaskTest extends AbstractTaskTest {
 
     def "exec npm task (download) with configured proxy"() {
         given:
-        props.setProperty('os.name', 'Linux')
         nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
         nodeExtension.download.set(true)
         def nodeDir = projectDir.toPath().resolve(".gradle").resolve("nodejs")
@@ -103,7 +96,6 @@ class NpmTaskTest extends AbstractTaskTest {
         GradleProxyHelper.setHttpProxyPort(123)
 
         def task = project.tasks.create('simple', NpmTask)
-        mockPlatformHelper(task)
         mockProjectApiHelperExec(task)
         task.args.set(["run", "command"])
 
@@ -128,13 +120,11 @@ class NpmTaskTest extends AbstractTaskTest {
 
     def "exec npm task with configured proxy"() {
         given:
-        props.setProperty('os.name', 'Linux')
         nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
         GradleProxyHelper.setHttpsProxyHost("my-super-proxy.net")
         GradleProxyHelper.setHttpsProxyPort(11235)
 
         def task = project.tasks.create('simple', NpmTask)
-        mockPlatformHelper(task)
         mockProjectApiHelperExec(task)
         task.args.set(['a', 'b'])
 
@@ -150,14 +140,12 @@ class NpmTaskTest extends AbstractTaskTest {
 
     def "exec npm task with configured proxy but disabled"() {
         given:
-        props.setProperty('os.name', 'Linux')
         nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
         GradleProxyHelper.setHttpsProxyHost("my-super-proxy.net")
         GradleProxyHelper.setHttpsProxyPort(11235)
         nodeExtension.nodeProxySettings.set(ProxySettings.OFF)
 
         def task = project.tasks.create('simple', NpmTask)
-        mockPlatformHelper(task)
         mockProjectApiHelperExec(task)
         task.args.set(['a', 'b'])