+++ src/test/groovy/com/github/gradle/node/npm/task/NpxTaskTest.groovy
@@ -4,15 +4,14 @@ package com.github.gradle.node.npm.task
 import com.github.gradle.node.task.AbstractTaskTest
 import com.github.gradle.node.util.PlatformHelperKt
 import com.github.gradle.node.variant.VariantComputer
+import com.github.gradle.node.variant.VariantComputerKt
 
 class NpxTaskTest extends AbstractTaskTest {
     def "exec npx task"() {
         given:
-        props.setProperty('os.name', 'Linux')
         nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
 
         def task = project.tasks.create('simple', NpxTask)
-        mockPlatformHelper(task)
         mockProjectApiHelperExec(task)
         task.command.set('command')
         task.args.set(['a', 'b'])
@@ -36,11 +35,9 @@ class NpxTaskTest extends AbstractTaskTest {
 
     def "exec npx task (windows)"() {
         given:
-        props.setProperty('os.name', 'Windows')
         nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Windows", "x86_64", {}))
 
         def task = project.tasks.create('simple', NpxTask)
-        mockPlatformHelper(task)
         mockProjectApiHelperExec(task)
         task.command.set('command')
         task.args.set(['a', 'b'])
@@ -64,16 +61,14 @@ class NpxTaskTest extends AbstractTaskTest {
 
     def "exec npx task (download)"() {
         given:
-        props.setProperty('os.name', 'Linux')
         nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
         nodeExtension.download.set(true)
-        def variantComputer = new VariantComputer(testPlatformHelper)
+        def variantComputer = new VariantComputer()
         def nodeDir = nodeExtension.resolvedNodeDir
-        def nodeBinDir = variantComputer.computeNodeBinDir(nodeDir)
-        def npxScriptFile = variantComputer.computeNpmScriptFile(nodeDir, "npx")
+        def nodeBinDir = variantComputer.computeNodeBinDir(nodeDir, nodeExtension.resolvedPlatform)
+        def npxScriptFile = VariantComputerKt.computeNpmScriptFile(nodeDir, "npx", nodeExtension.resolvedPlatform.get().isWindows())
 
         def task = project.tasks.create('simple', NpxTask)
-        mockPlatformHelper(task)
         mockProjectApiHelperExec(task)
 
         when: