+++ src/test/groovy/com/github/gradle/node/pnpm/task/PnpmTaskTest.groovy
@@ -6,11 +6,9 @@ import com.github.gradle.node.util.PlatformHelperKt
 class PnpmTaskTest extends AbstractTaskTest {
     def "exec pnpm task"() {
         given:
-        props.setProperty('os.name', 'Linux')
         nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
 
         def task = project.tasks.create('simple', PnpmTask)
-        mockPlatformHelper(task)
         mockProjectApiHelperExec(task)
         task.args.set(['a', 'b'])
         task.environment.set(['a': '1'])
@@ -31,11 +29,9 @@ class PnpmTaskTest extends AbstractTaskTest {
 
     def "exec pnpm task (windows)"() {
         given:
-        props.setProperty('os.name', 'Windows')
         nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Windows", "x86_64", {}))
 
         def task = project.tasks.create('simple', PnpmTask)
-        mockPlatformHelper(task)
         mockProjectApiHelperExec(task)
         task.args.set(['a', 'b'])
         task.environment.set(['a': '1'])
@@ -56,12 +52,10 @@ class PnpmTaskTest extends AbstractTaskTest {
 
     def "exec pnpm task (download)"() {
         given:
-        props.setProperty('os.name', 'Linux')
         nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
         nodeExtension.download.set(true)
 
         def task = project.tasks.create('simple', PnpmTask)
-        mockPlatformHelper(task)
         mockProjectApiHelperExec(task)
 
         when: